% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absoluteConnections.R
\name{absoluteConnections}
\alias{absoluteConnections}
\title{Determine the absolute Number of Connections (nC) for all patches}
\usage{
absoluteConnections(habitatraster, kernel, minp = 0.05, summarize = TRUE)
}
\arguments{
\item{habitatraster}{Raster containing patches of suitable habitats, coded with values > 0.}

\item{kernel}{Dispersal kernel of the species.}

\item{minp}{Minimum fraction of eC at which two patches are connected.}
}
\value{
A table or raster with the number of connections (nC) for each Patch
}
\description{
Determine the absolute Number of Connections (nC) for all patches
}
\examples{
data(habitats_lech)
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)
eC <- absoluteConnections(habitats_lech, sddkernel_chondrilla, minp=0.01)

}
