% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colonizationPotential.R
\name{colonizationPotential}
\alias{colonizationPotential}
\title{Determine the colonizationPotential (cP) for all patches}
\usage{
colonizationPotential(habitats, kernel, cap = FALSE, summarize = TRUE)
}
\arguments{
\item{habitats}{Raster containing patches of suitable habitats, coded with values > 0.}

\item{kernel}{Dispersal kernel of the species.}

\item{cap}{If true, values larger than 1 are capped.}

\item{summarize}{If TRUE, a table instead of a raster is returned.}
}
\value{
A table or raster with the colonizationPotential for each Patch
}
\description{
Determine the colonizationPotential (cP) for all patches
}
\examples{
example <- system.file("extdata/cc_clumped_lech_ehd.tif", package = "rconnect")
habitats_lech <- raster::raster(example)
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)
cP <- colonizationPotential(habitats_lech, sddkernel_chondrilla, cap=TRUE)

}
