% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersalKernel.R
\name{dispersalKernel}
\alias{dispersalKernel}
\title{Create a rectangular dispersal kernel}
\usage{
dispersalKernel(
  cellsize = 1,
  radius = 3,
  decay = 1,
  centervalue = 0,
  weighted = T
)
}
\arguments{
\item{cellsize}{Cell size of your habitat raster.}

\item{radius}{Radius of the kernel in cells.}

\item{decay}{Decay of the negative exponential function of the kernel.}

\item{centervalue}{value that should be set for the focal cell.}

\item{weighted}{Wether the kernel should be normalized}
}
\value{
A matrix with the kernel probablity density.
}
\description{
Create a rectangular dispersal kernel
}
\examples{
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)

}
