% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveDistance.R
\name{effectiveDistance}
\alias{effectiveDistance}
\title{Calculate effectiveDistance (eD) for habitat patches}
\usage{
effectiveDistance(habitats, kernel, summarize = TRUE)
}
\arguments{
\item{habitats}{Raster with the effective seedrain, summed up for each patch.}

\item{kernel}{kernel Dispersal kernel of the species.}

\item{summarize}{If TRUE, a table instead of a raster is returned.}
}
\value{
A raster or a summary table with the effective Distance (eD) for each Patch
}
\description{
Calculate effectiveDistance (eD) for habitat patches
}
\examples{
example <- system.file("extdata/cc_clumped_lech_ehd.tif", package = "rconnect")
habitats_lech <- raster::raster(example)
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)
eD <- effectiveDistance(habitats_lech, sddkernel_chondrilla)

}
