% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveSeedrain.R
\name{effectiveSeedrain}
\alias{effectiveSeedrain}
\title{Determines the effective Seed Rain (eS)}
\usage{
effectiveSeedrain(habitats, kernel, summarize = TRUE, fun = NA)
}
\arguments{
\item{habitats}{Raster containing patches of suitable habitats, coded with values > 0.}

\item{kernel}{Dispersal kernel of the species.}

\item{summarize}{If true, a summary table will be returned instead of a raster.}

\item{fun}{A function used for summary. If summarize=TRUE and none is provided sum is used.
If summarize is FALSE and a function is provided, a zoal statistics is applied to the raster}
}
\value{
A raster or a table with the effective seed rain (eS)
}
\description{
Determines the effective Seed Rain (eS)
}
\examples{
data(habitats_lech)
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)
eS <- effectiveSeedrain(habitats_lech, sddkernel_chondrilla, fun=sum)

}
