% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitats_lech}
\alias{habitats_lech}
\title{Raster of a riverscape at the Tirolean Lech}
\format{
UTM32 Raster data with 120 columns and 121 rows, cell size 5m,
\describe{
\item{value}{Patch ID}
}
}
\usage{
habitats_lech
}
\description{
A raster containing suitable habitat patches in a riverscape at the Lech.
}
\keyword{datasets}
