% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absoluteConnections.R
\name{absoluteConnections}
\alias{absoluteConnections}
\title{Determine the absolute Number of Connections (nC) for all patches}
\usage{
absoluteConnections(habitats, kernel, threshold = 0.05, summarize = TRUE)
}
\arguments{
\item{habitats}{Raster containing patches of suitable habitats, coded with values > 0.}

\item{kernel}{Dispersal kernel of the species.}

\item{threshold}{Minimum fraction of eC at which two patches are connected.}

\item{summarize}{If TRUE, a table instead of a raster is returned.}
}
\value{
A table or raster with the number of connections (nC) for each Patch
}
\description{
Determine the absolute Number of Connections (nC) for all patches
}
\examples{
example <- system.file("extdata/cc_clumped_lech_ehd.tif", package = "rconnect")
habitats_lech <- raster::raster(example)
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)
nC <- absoluteConnections(habitats_lech, sddkernel_chondrilla, threshold=0.01)


}
