% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colonizationPotential.R
\name{colonizationPotential}
\alias{colonizationPotential}
\title{Summarize the colonizationPotential (cP) for a riverscape}
\usage{
colonizationPotential(
  habitats,
  kernel,
  threshold = 0,
  ed_with_threshold = FALSE,
  cap = FALSE
)
}
\arguments{
\item{habitats}{Raster containing patches of suitable habitats, coded with values > 0.}

\item{kernel}{Dispersal kernel of the species.}

\item{threshold}{Minimum of eC at which two patches are connected.}

\item{ed_with_threshold}{If true, only patches within threshold are considered for eDm.}

\item{cap}{If true, values larger than 1 are capped.}
}
\value{
A table or raster with the colonizationPotential for each Patch
}
\description{
Summarize the colonizationPotential (cP) for a riverscape
}
\examples{
example <- system.file("extdata/cc_clumped_lech_ehd.tif", package = "rconnect")
habitats_lech <- raster::raster(example)
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)
cP <- colonizationPotential(habitats_lech, sddkernel_chondrilla, threshold=0.01, cap=TRUE)

}
