% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveConnections.R
\name{effectiveConnections}
\alias{effectiveConnections}
\title{Determine the effective Connections (eC) for all patches}
\usage{
effectiveConnections(
  habitats,
  kernel,
  threshold = 0,
  weighted = TRUE,
  cap = FALSE,
  summarize = TRUE
)
}
\arguments{
\item{habitats}{Raster containing patches of suitable habitats, coded with values > 0.}

\item{kernel}{Dispersal kernel of the species.}

\item{threshold}{Minimum of eC at which two patches are connected.}

\item{weighted}{If false, the number of connection is returned.}

\item{cap}{If true, values larger than 1 are capped.}

\item{summarize}{If TRUE, a table instead of a raster is returned.}
}
\value{
A table or raster with the number of connections for each Patch
}
\description{
Determine the effective Connections (eC) for all patches
}
\examples{
example <- system.file("extdata/cc_clumped_lech_ehd.tif", package = "rconnect")
habitats_lech <- raster::raster(example)
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)
eC <- effectiveConnections(habitats_lech, sddkernel_chondrilla, threshold=0.01, cap=TRUE)

}
