% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveConnectionsMatrix.R
\name{effectiveConnectionsMatrix}
\alias{effectiveConnectionsMatrix}
\title{Returns a matrix with the effective Connections (eC) between patches}
\usage{
effectiveConnectionsMatrix(habitats, kernel, threshold = 0.05)
}
\arguments{
\item{habitats}{Raster containing patches of suitable habitats, coded with values > 0.}

\item{kernel}{Dispersal kernel of the species.}

\item{threshold}{Minimum of eC at which two patches are connected.}
}
\value{
A matrix with patch-to-patch effectiveConnectivity
}
\description{
Returns a matrix with the effective Connections (eC) between patches
}
\examples{
example <- system.file("extdata/cc_clumped_lech_ehd.tif", package = "rconnect")
habitats_lech <- raster::raster(example)
sddkernel_chondrilla <- dispersalKernel(cellsize=5, radius=3, decay=0.19)
eCM <- effectiveConnectionsMatrix(habitats_lech, sddkernel_chondrilla, threshold=0.01)

}
